requirements.txt
fastapi
uvicorn
pdfplumber
gspread
google-auth
python-multipart
